Move utf8ness calc for $! into locale.c from mg.c
locale.c has the infrastructure to handle this, so remove repeated
logic.

The removed code tried to discern better based on using script runs, but
this actually doesn't help, so is removed.

Since we're now using C99, we can remove the block that was previously
needed, and now the code is properly indented, whereas before it wasn't