Add locale unsupported test
Perl only suppots multi-byte locales that are UTF-8.  It turns out that
the others are worse than I thought, and if someone switches to one, the
program can crash.

This commit generates a default-on diagnostic when switching into such a
locale.

The check has been done in various releases for some time, but this
elevates its severity.

Commit 74ecb5a150c08c23c0219dcaf8c55af7ba696bd5 caused the test suite to
view a locale that generates this message as unacceptable.  The commit
message for that commit said the next commit would be the one generating
the diagnostic.   Due to rebasing changes, that didn't happen until this
current commit.