cygwin.c: Add utf8_to_wide_extra_len()
This function is like utf8_to_wide, but has an additional parameter that
reserves extra space.  It thus is a superset of utf8_to_wide() which
changes to use it with 0 extra space needed.

And there is other code that parallels utf8_to_wide() but needs extra
space reserved.  That code is converted to use this (new, now common)
function, thus simplifying things.