INSTALL now says that "changes in the Perl language" are usually minor
The paragraph starts:

    On a related issue, old modules may possibly be affected by the changes
    in the Perl language in the current release.

Previously it made no indication about whether these changes were frequent
or rare, small or large. The reality is that we are careful to keep them
small and backwards compatible, whenever possible. So add the text:

    We try hard to make new features "opt-in", such that existing code will
    work unchanged, and attempt to identify where bug fixes might expose
    code which was relying on incorrect interpreter behaviour.

Also remind people that the pod/perldelta.pod is only changes for the
current release, hence they should read all the relevant perldelta files
if they are upgrading from a release further back. perldelta already says
this, but INSTALL was not clear.