Replace "define\t" with "define " in Configure/metaconfig related files
This is broken out so it is easier for Tux to find and merge with
metaconfig.

View this patch with -w and you will see "no changes" except for
config_h.SH and Porting/config_h.pl both which needed to be changed to
ensure that they produce output that doesn't replicate the problem.