bisect-runner.pl must patch Configure to avoid myread looping infinitely
Configure can get stuck and ask questions for which it needs a valid answer
before it can continue. As-is, if you redirect stdin from /dev/null (or close
the file descriptor) it will (effectively) loop infinitely repeating the same
question, because it doesn't like empty string as an answer. Worse - it keeps
repeating the question to stdout - eg 'Where is your C library?'

Rather than attempting to patch the shell script to detect errors on read
(because they only matter the *second* time round the loop, *and* wouldn't
handle the /dev/null case), it's easier to patch the relevant loop so that
it will abort after too many loop iterations.