bisect-runner.pl needs to avoid using nm on darwin
Earlier versions of the hints defaulted to using nm, because on older versions
of OS X (as was) it worked.

It also needs to patch the hints file to force d_stdstdio to "undef".
Without this, a build with "d_faststdio" defined (or implicit) will fail badly
on current macOS, such as the perl-5.8.0 tag.