bisect-runner.pl needs to force a -flat_namespace before v5.10.0
The hints for macOS set -flat_namespace conditionally based on darwin
version, so that newer OSes would default to the native two level namespace.
However, the build of miniperl was relying on a flat namespace prior to a
refactoring during the 5.9.x series. Hence we need to force this linker flag
when building versions before this on current macOS versions.