t/io/eintr.t: Fix sporadic test failure
The test would sporadically fail; troubleshooting this revealed that
the size of the pipe buffer could vary between pipes which this code
didn't take into account.

What the code did:
	1. Create two pipes ($in, $out)
	2. Get buffer size of '$out' pipe
	3. Do something that should block (using the buffer size)
	4. Create two new pipes ($in, $out)
	5. Do something that should block (using the buffer size)

The problem: the pipe created in 4. does not always have the same
             buffer size as the pipe created in 1.

Adding some debug info in the test:
	# pipe buf size: 8193
	ok 4 - read/die: read status
	ok 5 - read/die: close status
	# pipe size: 65537
	not ok 6 - print/close: sig handler close status

During test 6 it reused the smaller buffer size and wrote data
to the new pipe. It expected this to block but the buffer of the pipe
used for test 6 is 65536 bytes. End result: the pipe buffer is not
full (only 8194 bytes were written) so the test does not block which
causes the alarm not to trigger which causes the test failure.

Fix all this by reinitialising `$pipe_buf_size` when creating new pipes.

(Observed by @khwilliamson, diagnosed with help of @wolfsage)