Handle intrin files on win32 with gcc
This fixes #20033.

When building on Windows with Strawberry 5.32.1 (gcc 8.3.0) as the toolchain,
the Errno.pm is created by a script Errno_pm.pl, which takes output from the
compiler to find headers.

A subset of these headers requires them to only be included by some specific
headers. Previously the header order was effectively random and this
occasionally caused build errors (that further were never detected).

The get_files() is now returning the header names in the order the compiler
saw them which insures they are in the right order.