diag.t: subDivide exception list
Some of the entries in this list of diagnostics are ones that people
just haven't gotten around to documenting.

Patches welcome!

But there are, in my opinion, a very few are ones that don't ever need
to be documented because the text of the diganostic is sufficiently
explanatory in and of itself.

This commit revises the comments explaining the exception list, and
moves the ones I think are sufficiently self-explanatory to the front.