CI: Remove fetching of tags from sanity check
For the 'sanity check' CI the option `fetch-depth: 0` is used.
This causes it to fetch all branches *and* all tags with all history.

By doing `git fetch --depth=1 origin +refs/tags/*:refs/tags/*` it is
fetching all the tags again and in this case *discarding* all the history
of the tags and turning this into a shallow clone instead of a full clone.

-> Remove the fetching of tags from the sanity check CI since they were
   already fetched with their full history.

(This is needed because in the next commit a check will be done to see
 if the test is running in a shallow clone)