Consolidate PERL_TEST_TIME_OUT_FACTOR to watchdog()
This changes test.pl watchdog() to always consider this potential
setting of an environment variable, and removes the distributed uses.

This means that no code needs to change when tests start failing on a
slow platform due to timing out; or when you need to temporarily
increase the timeout of a test for debugging.  This will correspondingly
increase the timeout of all tests, but who cares for debugging purposes.