retainedlines.t - deterministic results and fixup tests under failure
The existing logic is only "correct" when all the tests pass. When they
fail the tests are revealed to be non-determinstic. The other problem is
that once a single test case fails and leaks an entry in the stash the
following tests are all contaminated and trigger "false failures". The
combination makes things look much more broken than they are.

This patch fixes the tests so that they are deterministic and the
"leaking" between test cases is stopped.

See GH Issue #20174