Move '=encoding utf8' after preamble
All the READMEs start with a preamble describing that the file is
written in the POD format.

When the '=encoding' line is put before the preamble then the
preamble is also rendered when viewing the file with `perldoc`
(and/or when viewing the file as POD).

Fix: move '=encoding' to the place where the actual POD starts.