pad.h - reword comment to be more legible
The prevous verbiage ended the comment with the term "padnamelist *",
which put the "*" right next to the end comment marker, which was
a bit confusing. This wording put the * in the middle of the sentence
where it is clear it is not a mistake.

Also note that the previous text used tabs internally, and so this change
/looks/ like it is off indent, it is not, it is lined up with the surrounding
lines.