testsuite.yml: Remove fetching of tags
The only test that need the git tags (afaik) is porting/cmp_version.t;
But the fetching of tags isn't really working for that test.

Basically the repo is cloned with:
	$ git clone --depth 1 https://github.com/perl/perl5
	$ git fetch --depth=1 origin +refs/tags/*:refs/tags/*

And then later the test uses:
	$ git describe --abbrev=0 HEAD

and it expects to find a tag.

However doing that:
	$ git describe --abbrev=0 HEAD
	fatal: No tags can describe '6a011f13d7690dbe2e03ad7500756c983bcb1834'.

So it's not really working..

This can also be seen in the CI logs[^1], the output contains:
	porting/cmp_version.t ..... skipped: Git found, but no Git tags found

There are only 3 CI configs that do run the porting/cmp_version.t test:
	- 'Sanity: Linux -Dusethreads' - this is intended
	- 'linux i386/ubuntu (-Dusedevel)'[^2]
	- 'cygwin'[^3]

[^1]: CI logs of: 'linux (-Uusethreads)', 'linux (-Duseshrplib ...)',
      'linux (-Duserelocatableinc)', 'linux (-Dcc='clang')', 'linux (-Dcc='g++')',
      'install', 'macOS (catalina) xcode 12 (-Uusethreads)',
      'macOS (catalina) xcode 12 (-Duseithreads -Duseshrplib)', 'Windows msvc142',
      'Windows mingw64', 'ASAN (-Accflags=-DPURIFY)', 'ASAN (-DDEBUGGING)',
      'ASAN (-Dusethreads)', 'PERL_UNICODE (-Uusethreads)',
      'PERL_UNICODE (-Dusethreads ...')

[^2]: it uses 'actions/checkout@v1' to do the checkout which does a full clone
[^3]: it manually runs `git clone`, doing a full clone (with all history)