mktables: Accept multiple @missing lines in input files
Unicode 15.0 will now use this approach to deal with ranges of code
points that have a different default for unassigned code points than the
table at large.  For example, a table may have one default, but all
Ideographic character ranges have something else.  Prior to this new
mechanism, the files had entries for each unassigned code point that had
a different default than the global one.  So this saves some lines in
the files that Unicode delivers that were otherwise useless.

Not all files in 15.0 have been converted to use the new scheme, for
whatever reason.