mktables: Skip some new 15.0 files
These are newly delivered by Unicode.  I haven't had time to analyze
them for use for potential new properties.  They deal with security
issues of characters that look alike.

I'm not adding them to the list of files under git, but they are
explicitly mentioned in mktables to indicate their not being used.