Extend cmpVERSION.pl to check for modified .c/.h
When a .c or .h files is modified in ext/, cpan/, dist/ then check
if another file in that dist was bumped.

The inspiration for this extra check: sometimes a .c or .h file gets
modified without bumping a version number in any of the other files.
This then causes two things to report the same version number while
they are in fact somewhat different (assuming the .c/.h code is used).

This seems less optimal so add an extra check: if a .c/.h file is
modified then something in that dist should also have a version-bump.