mktables: Standardize value aliases
As stated in the code comments added by this commit, Unicode has various
spellings for the same property value.  For example in some places it
uses 'W', and in others 'Wide'.  The legal spellings are listed in
PropValueAliases.txt, which is processed early in the construction.  So
we can standardize things on input, which makes it easier later.  This
commit produces minimal changes in the generated tables, so that the
algorithm can be verified by inspection of the results.  And no other
code that has hard-coded in expected spellings needs to be changed.

Prior to this commit, we standardized the default value for properties
that have a default value,.