mktables: Convert array to hash
Prior to this commit we had a two element array, and it was known that
element 0 contained a particular thing; and element 1 contained the
other.  But a future commit will add several elements, so keeping track
of which is which will become more problematic.  Solve this by using a
hash instead, with the elements appropriately named.