CI/dist-modules: add threads.h for threads/threads-shared
threads and threads-shared both bundle this on CPAN.  It provides
empty definitions for a few macros that may not have been defined in
ppport.h.  For now this includes the version from CPAN, but we may be
able to optimize it to a smaller version given that ppport.h has been
updated.

Since we now have two files, rather than including them as (fairly
large for t/test.pl) here-docs, move the content to __DATA__.  I've
retained the individual functions that forward onto _write_from_data()
to provide a place to document why those files are needed.