pat_advanced.t: Cancel a watchdog timer
This timer had been skipped until
91d110e02db3b8905c4c48d5fdd76a8439de31f3.  But with it enabled, on
slower or heavily loaded systems the test file does not always complete.
Take advantage of the new ability to cancel watchdog timers to prevent
this situation.

Another possibility would to have been to move this test to be the final
one in the file.  But then effort has to be expended to keep it in that
position as new tests are added; and what if such tests also require a
watchdog?  It's simplest to have the watchdog affect only the single
test that needs it.

Yet another possibility is to make this test into fresh_perl.  But that
sort of interferes with what t/re/pat_advanced_thr.t does.