Remove sharedperl option on VMS
This was intended to allow Perl for multiple architectures to reside
in a single installation directory tree, primarily for mixed-architecture
clusters.  Disk space is a lot cheaper than it was when this first came
along, and architecture-specific library directories have since become
a thing, so there might be a better way to do this now if it's even still
desireable.  It's unclear whether the feature has been used in recent
decades -- it certainly hasn't been tested recently.

Removing it simplifies the build and is something that doesn't have to
get fixed whenever a port to a new architecture comes along.