pp_ctl.c - in try_run_unitcheck() guard against leaking PL_restartop
If we die while executing a UNITCHECK inside of an eval we shouldn't
leave PL_restartop set, as we will execute PL_op->op_next anyway. See
the previous commit for more details in the context yy_parse().

Thanks to Bram for coming up with a test case that demonstrated the
problem.