Update Module::CoreList for 5.37.6
In 4ab96809c99e944e70c21779641e4b1c9a00df41 the perl version
data was bumped, without doing any Module-CoreList changes and
now related tests are failing. Release managers guide said to
do:

    ./perl -Ilib Porting/corelist.pl cpan

to update the corelist data, which is this commit. Not sure if
this is 100% the right thing to do, but it makes the test fails
stop for now. Someone who knows this better can do any mop-up.