Devel::PPPort/module3: eliminate the always true warning for the return
In this case we're testing that the result is always true, and it
happens the compiler realizes that the return value is always true,
even in isolation.

Returning the variable that we've already set to point to
PL_bufptr eliminates the warning.

The compiler is still smart enough to eliminate the comparison
on optimized builds.  Some linter may still complain about it.