locale.c: Don't Safefree temps
Commit 22392ba2f8a changed the querylocale_foo macros to return a
mortalized copy instead of having to free things explicitly.  This was
done because the various code paths are complicated, and I was noticing
various leaks that would require a bunch of #ifdef's to fix.

That commit meant that all the existing free's of the results needed to
be removed.  This one was overlooked at the time.  Reported in
<f47046d6-7e35-301d-10f-9160b34ca2b6@uwe-dueffert.de>, not in github