POSIX::localeconv: Return empty/special values
This function returns a hash allowing Perl access to the localeconv()
data structure, with the keys being the structure's field names, and the
values being their corresponding value in the current locale.

Prior to this commit, it did not populate the hash with:
    1) any string-valued keys whose value is the empty string
    2) any numeric-valued keys whose value is the special value CHAR_MAX

This is wrong.  localeconv() should return a complete list of fields on
the platform, regardless of their values.  Someone may well wish
to iterate over all the keys in the hash.

CHAR_MAX just indicates special handling is required for that
numeric field.  And all string fields legally can be empty, except for
the decimal point.  For example, the symbol indicating a number is
positive is empty in many locales.

I couldn't find a reason in the history why these have been omitted.