Avoid use of Vietnamese UTF-8 locale on Solaris
On OmniOS r151042 (based on Solaris) the 'strfxm_l' function segfaults
when:

- LC_COLLATE_MASK is set to 'vi_VN.UTF-8'
- the input contains "\xA8"

Other locales and other characters appear to be unaffected.

This commit causes 'loc_tools.pl' to not return this locale as being
acceptable when asked for a list of them.  Thus, our automated locale
testing will avoid it.

A Perl program can still switch into it explicitly.
This issue is tracked by #20578, which includes a C-only reproducer not
involving Perl.