locale.c: Create a convenience #define
This takes this #ifdef combination and creates a single symbol that
means essentially the same thing.  This is in preparation for it being
used in additional places.  And it makes it clearer what is meant.

It also adds another condition which isn't currently tested for, but has
the same meaning, and doesn't affect the current usage, but will future
ones.