perlsub: normalize whitespace to use 4 space indents
The file was previously using a mix of indent styles. Some parts used 4
space indents, some used 4 space indents mixed with tabs. The file would
not be properly viewable with a tab width other than 8. This is normal
for many files in perl core, but they should all be fixed.

Normalize all of the indents to use 4 space indents consistently, with
no tabs.