harness: Run more parallel tests by execution time order
In the normal mode of operation of t/harness, there are three sets of
tests, where the first set has to completely finish before the second
one starts, which has to completely finish before the third set starts.

The first set is the basic sanity, and quickly completes.  That is
untouched by this commit.

The third set are the tests for mostly non-core modules, including
cpan/, dist/, ext/, and lib/.  These are executed so that the
longest-running ones are started first when parallel testing is
available.  That is also unchanged by this commit.

The second set is what's left over: generally core tests and ones
enabled only on some platforms or by specifying an option, such as
torture tests.  Prior to this commit, these were executed in parallel,
if permitted, but there was no effort to start the longest-running ones
first.  This results in a less than optimum use of the CPU.

The code that orders tests by running time was previously applied to
just the third set of tests.  Prior commits have extracted out that code
into a function.  This commit just uses that function for the tests
in the second set.

In principle, the basic sanity tests could also be run in parallel.
This would be easy to do, but I'm a little leery, and since they are
quick to execute there would not be much gain in doing so.

Comments in the code removed by this commit indicated that the intent
was to execute ext/ and lib/ tests last.  But that hasn't been true for
quite some time, and no one seems to have noticed.  Now that we have a
function to assemble the tests in a given directory, it would be trivial
to change back, but that does lead to a longer wall clock elapsed time.

The PERL_TEST_HARNESS_ASAP environment variable is used to change the
mode of operation so that most of the tests in the second set are
instead moved into the third set.  Only the comp/, run/, and cmd tests
are left in the second set.  That is also unchanged by this commit.