harness: When run in ASAP mode, intermix comp/run/cmd
Prior to this commit, when run with PERL_TEST_HARNESS_ASAP, there were
three sets of tests run sequentially.  The second set consisted of just
the comp/, run/, and cmd/ directories.  This commit folds those
directories and the third set together, which ekes out a bit more
performance gain.  It was 6% better on my box running ASAN.  And, the
ASAP runs in about 75% of the time as non-ASAP.

There is no change in the non-ASAP operation.