API-test:locale.t: Look for a comma radix locale
Prior to this commit there was code (whose results were ignored) looking
for a locale with a non-dot radix.  This can result in a UTF-8 radix,
which may not display properly without the terminal and file handle
being in sync.  Almost all non-dot locales use a comma, which is
represented the same in UTF-8 as not, so doesn't suffer from the display
problem.  So look specifically for a comma.

The result is still unused, but the next commit will use it.