Move my_strftime() from util.c to locale.c
Stress testing has shown that unless LC_TIME and LC_CTYPE are the same
locale, mojibake can result on some platforms.

locale.c already handles this for various other similar cases, and has
the infrastructure available to easily handle this one too.  util.c does
not.

The only code change is to a comment to reflect that things are now in a
different file.