File-Find/t/taint.t - do not use rel2abs after we have forced to Unix mode
taint.t corrupts @INC, by using rel2abs twice on its values, once after
we have forced File::Spec to think we are on a *nix style platform. This
breaks use statements that occur after the second rel2abs. Comparing to
File-Find/t/find.t which also does this we can see that find.t only does
this rel2abs on @INC behavior once. Presumably the second call was a
rebase error of some sort.

This reorganizes the use statements and ensures that we only use
rel2abs() to change @INC prior to overriding it to use *nix semantics.