regen/embed_lib.pl - don't add integer to reference and fixup source data
Dave noticed there was a bug where I was adding a reference to an
integer. That did not make any sense. This patch fixes up the data so
the source of the line is more clear. I will do a further follow up
to improve this more, this is just to get the bug out of the way.