snprintf() calls need to have proper radix
Calls to libc snprintf() were neglected to be changed when perl was
fixed to change the radix character to the proper one based on whether
or not 'use locale' is in effect.  Perl-level code is unaffected, but
core and XS code is.

This commit changes to wrap snprintf() calls with the macros designed
for the purpose, long used for similar situations elsewhere in the code.

Doing this requires the thread context.  I achieved this in a few places
by a dTHX, instead of assuming a caller would have the context already
available, and adding a pTHX_ parameter.  I tried doing it the other
way, and got a few breakages in our test suite.  Formatting already
requires significant CPU time, so this addition should just be in the
noise

This bug was found by new tests that will be added in a future commit.