perly.* - update generated files to include file type in modeline
This is a companion patch to the previous one, it is generated with a
slightly more modern Bison which produces better indented output, and
puts parens around constant defines and other minor changes. It is
recommended to view this patch with the -w option to skip most of the
the "non change" changes it contains.