warnings.pm - support deprecated::smartmatch category
Currently we seem to lack a way to have a subcategory under deprecated.
It seems reasonable to me that people might want to disable a specific
subcategory warning while leaving the rest in place. This patch allows
that. Note that both

    no warnings "deprecated";

and

    no warnings "deprecated::smartmatch";

work to disable the warning. Deprecated warnings shouldn't be "all or
nothing", they should be specific and targetted.