mg.c: Split out functionality to a seprate function
This commit creates a function that gets the extended errno on systems
(such as VMS and Windows) that have that, returning plain errno on
systems that don't.  Previously this code was interweaved into code with
a larger purpose.  The splitting makes it possible for other code that
just wants the errno to get it.

The new function is callable from elsewhere in core, but remains private
for now.  In principal it could be documented and made public if needed.