Porting/cmpVERSION.pl - handle missing tags more gracefully
Today Mauke had a surprise where he was on the latest commit but
cmp_version.t kept failing. It turned out that his personal github
clone was missing the most recent tags, and thus confusing the test.

Since we now release a new tag monthly we can detect that this is
happening and skip the test with some helpful diagnositcs on how
to fix it.